Fix one more compiler warning missed by the previous check-in.
FossilOrigin-Name: bc03d99a78e90c02b69037e5f5f81537b5a3ac60

+++ ext/fts3/fts3.c
@@ -3990,7 +3990,7 @@ static int fts3EvalPhraseNext(
       }
       pDL->pList = pIter;
       fts3PoslistCopy(0, &pIter);
-      pDL->nList = (pIter - pDL->pList);
+      pDL->nList = (int)(pIter - pDL->pList);
 
       /* pIter now points just past the 0x00 that terminates the position-
       ** list for document pDL->iDocid. However, if this position-list was
